refactor(settings): decompose settings page into 8 focused sub-components
- Created src/components/settings/ with 8 section components:
  - AppearanceSettings.tsx (159L) - theme, colors, fonts, preview/save
  - CustomizerSettings.tsx (69L) - armor diagram mode, preferences
  - VaultSettings.tsx (24L) - vault identity, preferences
  - P2PSyncSettings.tsx (244L) - sync room, peer list, room code
  - UIBehaviorSettings.tsx (58L) - UI behavior toggles
  - AccessibilitySettings.tsx (45L) - accessibility preferences
  - AuditSettings.tsx (98L) - audit configuration
  - ResetSettings.tsx (48L) - reset/danger zone
- Reduced src/pages/settings.tsx from 1,354L to 193L (thin shell)
- All 26 characterization tests passing
- Largest component: 244L (< 400L limit)
- All components: 80-400L range (no micro-components)
