+++ src/pages/settings.tsx
@@ -10,834 +10,31 @@
  */
 
 import Head from 'next/head';
-import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React, { useEffect, useState, useRef, useCallback } from 'react';
 
-import { ArmorDiagramModePreview } from '@/components/customizer/armor/ArmorDiagramPreview';
-import { ArmorDiagramSettings } from '@/components/customizer/armor/ArmorDiagramSettings';
 import {
-  SyncStatusIndicator,
-  PeerList,
-  RoomCodeDialog,
-} from '@/components/sync';
-import { PageLayout, Input } from '@/components/ui';
-import { VaultIdentitySection } from '@/components/vault/VaultIdentitySection';
-import { ConnectionState } from '@/lib/p2p';
-import {
-  useSyncRoomStore,
-  useConnectionState,
-  usePeers,
-  useRoomCode,
-} from '@/lib/p2p/useSyncRoomStore';
-import {
-  useAppSettingsStore,
-  AccentColor,
-  FontSize,
-  AnimationLevel,
-  UITheme,
-  ACCENT_COLOR_CSS,
-} from '@/stores/useAppSettingsStore';
-
-/**
- * Section configuration for navigation and state
- */
-type SectionId =
-  | 'appearance'
-  | 'customizer'
-  | 'vault'
-  | 'p2p-sync'
-  | 'ui-behavior'
-  | 'accessibility'
-  | 'audit'
-  | 'reset';
-
-interface SectionConfig {
-  id: SectionId;
-  title: string;
-  description?: string;
-  icon: React.ReactNode;
-}
-
-const SECTIONS: SectionConfig[] = [
-  {
-    id: 'appearance',
-    title: 'Appearance',
-    description:
-      'Customize colors, fonts, and visual effects. Changes preview instantly but require saving.',
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="h-4 w-4"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M4.098 19.902a3.75 3.75 0 005.304 0l6.401-6.402M6.75 21A3.75 3.75 0 013 17.25V4.125C3 3.504 3.504 3 4.125 3h5.25c.621 0 1.125.504 1.125 1.125v4.072M6.75 21a3.75 3.75 0 003.75-3.75V8.197M6.75 21h13.125c.621 0 1.125-.504 1.125-1.125v-5.25c0-.621-.504-1.125-1.125-1.125h-4.072M10.5 8.197l2.88-2.88c.438-.439 1.15-.439 1.59 0l3.712 3.713c.44.44.44 1.152 0 1.59l-2.879 2.88M6.75 17.25h.008v.008H6.75v-.008z"
-        />
-      </svg>
-    ),
-  },
-  {
-    id: 'customizer',
-    title: 'Customizer',
-    description: 'Configure the mech customizer interface',
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="h-4 w-4"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M11.42 15.17L17.25 21A2.652 2.652 0 0021 17.25l-5.877-5.877M11.42 15.17l2.496-3.03c.317-.384.74-.626 1.208-.766M11.42 15.17l-4.655 5.653a2.548 2.548 0 11-3.586-3.586l6.837-5.63m5.108-.233c.55-.164 1.163-.188 1.743-.14a4.5 4.5 0 004.486-6.336l-3.276 3.277a3.004 3.004 0 01-2.25-2.25l3.276-3.276a4.5 4.5 0 00-6.336 4.486c.091 1.076-.071 2.264-.904 2.95l-.102.085m-1.745 1.437L5.909 7.5H4.5L2.25 3.75l1.5-1.5L7.5 4.5v1.409l4.26 4.26m-1.745 1.437l1.745-1.437m6.615 8.206L15.75 15.75M4.867 19.125h.008v.008h-.008v-.008z"
-        />
-      </svg>
-    ),
-  },
-  {
-    id: 'vault',
-    title: 'Vault & Sharing',
-    description: 'Manage your vault identity for sharing content',
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="h-4 w-4"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M7.217 10.907a2.25 2.25 0 100 2.186m0-2.186c.18.324.283.696.283 1.093s-.103.77-.283 1.093m0-2.186l9.566-5.314m-9.566 7.5l9.566 5.314m0 0a2.25 2.25 0 103.935 2.186 2.25 2.25 0 00-3.935-2.186zm0-12.814a2.25 2.25 0 103.933-2.185 2.25 2.25 0 00-3.933 2.185z"
-        />
-      </svg>
-    ),
-  },
-  {
-    id: 'p2p-sync',
-    title: 'P2P Sync',
-    description:
-      'Real-time peer-to-peer synchronization for collaborative play',
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="h-4 w-4"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M8.288 15.038a5.25 5.25 0 017.424 0M5.106 11.856c3.807-3.808 9.98-3.808 13.788 0M1.924 8.674c5.565-5.565 14.587-5.565 20.152 0M12.53 18.22l-.53.53-.53-.53a.75.75 0 011.06 0z"
-        />
-      </svg>
-    ),
-  },
-  {
-    id: 'ui-behavior',
-    title: 'UI Behavior',
-    description: 'Control how the interface behaves',
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="h-4 w-4"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M10.5 6h9.75M10.5 6a1.5 1.5 0 11-3 0m3 0a1.5 1.5 0 10-3 0M3.75 6H7.5m3 12h9.75m-9.75 0a1.5 1.5 0 01-3 0m3 0a1.5 1.5 0 00-3 0m-3.75 0H7.5m9-6h3.75m-3.75 0a1.5 1.5 0 01-3 0m3 0a1.5 1.5 0 00-3 0m-9.75 0h9.75"
-        />
-      </svg>
-    ),
-  },
-  {
-    id: 'accessibility',
-    title: 'Accessibility',
-    description: 'Options for better accessibility',
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="h-4 w-4"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M2.036 12.322a1.012 1.012 0 010-.639C3.423 7.51 7.36 4.5 12 4.5c4.638 0 8.573 3.007 9.963 7.178.07.207.07.431 0 .639C20.577 16.49 16.64 19.5 12 19.5c-4.638 0-8.573-3.007-9.963-7.178z"
-        />
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"
-        />
-      </svg>
-    ),
-  },
-  {
-    id: 'audit',
-    title: 'Audit Log',
-    description: 'View event history and system logs',
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="h-4 w-4"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"
-        />
-      </svg>
-    ),
-  },
-  {
-    id: 'reset',
-    title: 'Reset',
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="h-4 w-4"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M16.023 9.348h4.992v-.001M2.985 19.644v-4.992m0 0h4.992m-4.993 0l3.181 3.183a8.25 8.25 0 0013.803-3.7M4.031 9.865a8.25 8.25 0 0113.803-3.7l3.181 3.182m0-4.991v4.99"
-        />
-      </svg>
-    ),
-  },
-];
-
-/**
- * Quick navigation tags using URL hash anchors
- */
-function QuickNavigation({
-  activeSection,
-  onNavigate,
-}: {
-  activeSection: SectionId | null;
-  onNavigate: (sectionId: SectionId) => void;
-}) {
-  return (
-    <div className="bg-surface-deep/95 border-border-theme-subtle sticky top-0 z-10 -mx-4 mb-4 border-b px-4 py-3 backdrop-blur-sm">
-      <div className="flex flex-wrap items-center gap-2">
-        <span className="text-text-theme-muted mr-1 text-xs">Jump to:</span>
-        {SECTIONS.map((section) => {
-          const isActive = activeSection === section.id;
-          return (
-            <a
-              key={section.id}
-              href={`#${section.id}`}
-              onClick={(e) => {
-                e.preventDefault();
-                onNavigate(section.id);
-              }}
-              className={`inline-flex items-center gap-1.5 rounded-md px-2.5 py-1 text-xs font-medium no-underline transition-all duration-150 ${
-                isActive
-                  ? 'bg-accent/20 text-accent border-accent/30 border'
-                  : 'bg-surface-raised/50 text-text-theme-secondary border-border-theme-subtle hover:bg-surface-raised hover:text-text-theme-primary hover:border-border-theme border'
-              } `}
-            >
-              {section.icon}
-              {section.title}
-            </a>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-/**
- * Collapsible settings section wrapper
- */
-function SettingsSection({
-  id,
-  title,
-  description,
-  isExpanded,
-  onToggle,
-  onRef,
-  children,
-}: {
-  id: SectionId;
-  title: string;
-  description?: string;
-  isExpanded: boolean;
-  onToggle: () => void;
-  onRef?: (el: HTMLDivElement | null) => void;
-  children: React.ReactNode;
-}) {
-  return (
-    <div
-      ref={onRef}
-      id={`section-${id}`}
-      className="bg-surface-base/50 border-border-theme-subtle overflow-hidden rounded-lg border transition-all duration-200"
-    >
-      {/* Header - always visible, clickable to toggle */}
-      <button
-        onClick={onToggle}
-        className="hover:bg-surface-raised/30 flex w-full items-center justify-between p-5 text-left transition-colors"
-      >
-        <div className="flex-1">
-          <h3 className="text-text-theme-primary text-lg font-semibold">
-            {title}
-          </h3>
-          {description && !isExpanded && (
-            <p className="text-text-theme-muted mt-0.5 line-clamp-1 text-sm">
-              {description}
-            </p>
-          )}
-        </div>
-        <div
-          className={`text-text-theme-secondary ml-4 transition-transform duration-200 ${isExpanded ? 'rotate-180' : ''}`}
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-            strokeWidth={2}
-            stroke="currentColor"
-            className="h-5 w-5"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              d="M19.5 8.25l-7.5 7.5-7.5-7.5"
-            />
-          </svg>
-        </div>
-      </button>
-
-      {/* Content - collapsible */}
-      <div
-        className={`transition-all duration-200 ease-in-out ${isExpanded ? 'max-h-[2000px] opacity-100' : 'max-h-0 overflow-hidden opacity-0'} `}
-      >
-        <div className="border-border-theme-subtle/50 border-t px-5 pb-5">
-          {description && (
-            <p className="text-text-theme-secondary mt-4 mb-4 text-sm">
-              {description}
-            </p>
-          )}
-          <div className="space-y-4">{children}</div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Toggle switch component
- * Rectangular tactical style with proper sizing math
- */
-function Toggle({
-  label,
-  description,
-  checked,
-  onChange,
-}: {
-  label: string;
-  description?: string;
-  checked: boolean;
-  onChange: (checked: boolean) => void;
-}) {
-  return (
-    <div className="flex items-center justify-between gap-4">
-      <div className="min-w-0 flex-1">
-        <div className="text-text-theme-primary text-sm font-medium">
-          {label}
-        </div>
-        {description && (
-          <div className="text-text-theme-secondary mt-0.5 text-xs">
-            {description}
-          </div>
-        )}
-      </div>
-      <button
-        type="button"
-        role="switch"
-        aria-checked={checked}
-        onClick={() => onChange(!checked)}
-        className={`focus:ring-accent focus:ring-offset-surface-base relative h-7 w-12 flex-shrink-0 cursor-pointer rounded-md border-2 transition-colors duration-200 ease-in-out focus:ring-2 focus:ring-offset-2 focus:outline-none ${
-          checked
-            ? 'bg-accent border-accent-hover'
-            : 'bg-surface-raised border-border-theme-strong hover:bg-surface-raised/80 hover:border-border-theme-strong'
-        } `}
-      >
-        <span
-          aria-hidden="true"
-          className={`absolute top-0.5 left-0.5 h-5 w-5 rounded border border-slate-300 bg-white shadow-md transition-transform duration-200 ease-in-out ${checked ? 'translate-x-5' : 'translate-x-0'} `}
-        />
-      </button>
-    </div>
-  );
-}
-
-/**
- * Select dropdown component
- */
-function Select<T extends string>({
-  label,
-  description,
-  value,
-  options,
-  onChange,
-}: {
-  label: string;
-  description?: string;
-  value: T;
-  options: { value: T; label: string; description?: string }[];
-  onChange: (value: T) => void;
-}) {
-  return (
-    <div>
-      <div className="text-text-theme-primary mb-1 text-sm font-medium">
-        {label}
-      </div>
-      {description && (
-        <div className="text-text-theme-secondary mb-2 text-xs">
-          {description}
-        </div>
-      )}
-      <select
-        value={value}
-        onChange={(e) => onChange(e.target.value as T)}
-        className="bg-surface-raised border-border-theme text-text-theme-primary focus:ring-accent w-full rounded-lg border px-3 py-2 text-sm focus:ring-2 focus:outline-none"
-      >
-        {options.map((opt) => (
-          <option key={opt.value} value={opt.value}>
-            {opt.label}
-          </option>
-        ))}
-      </select>
-    </div>
-  );
-}
-
-/**
- * Color picker for accent colors
- */
-function AccentColorPicker({
-  value,
-  onChange,
-}: {
-  value: AccentColor;
-  onChange: (color: AccentColor) => void;
-}) {
-  const colors: { value: AccentColor; label: string }[] = [
-    { value: 'amber', label: 'Amber' },
-    { value: 'cyan', label: 'Cyan' },
-    { value: 'emerald', label: 'Emerald' },
-    { value: 'rose', label: 'Rose' },
-    { value: 'violet', label: 'Violet' },
-    { value: 'blue', label: 'Blue' },
-  ];
-
-  return (
-    <div>
-      <div className="text-text-theme-primary mb-2 text-sm font-medium">
-        Accent Color
-      </div>
-      <div className="text-text-theme-secondary mb-3 text-xs">
-        Customize the highlight color throughout the app
-      </div>
-      <div className="flex flex-wrap gap-4">
-        {colors.map((color) => (
-          <button
-            key={color.value}
-            onClick={() => onChange(color.value)}
-            className={`focus:ring-accent focus:ring-offset-surface-base h-10 w-10 rounded-lg transition-all duration-150 focus:ring-2 focus:ring-offset-2 focus:outline-none ${
-              value === color.value
-                ? 'scale-125 border-2 border-white shadow-xl ring-4 ring-white/50'
-                : 'border-border-theme hover:border-border-theme-strong border-2 hover:scale-110'
-            }`}
-            style={{ backgroundColor: ACCENT_COLOR_CSS[color.value].primary }}
-            aria-label={color.label}
-            title={color.label}
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-/**
- * UI Theme info for display
- */
-const UI_THEME_INFO: Record<
-  UITheme,
-  { name: string; description: string; preview: string }
-> = {
-  default: {
-    name: 'Default',
-    description: 'Clean slate design with amber accents',
-    preview: 'bg-surface-base border-border-theme-subtle',
-  },
-  neon: {
-    name: 'Neon',
-    description: 'Cyberpunk-inspired with glow effects',
-    preview: 'bg-surface-deep border-cyan-500/30',
-  },
-  tactical: {
-    name: 'Tactical',
-    description: 'Military HUD style with monospace fonts',
-    preview: 'bg-surface-deep border-border-theme',
-  },
-  minimal: {
-    name: 'Minimal',
-    description: 'Reduced visual noise, focus on content',
-    preview: 'bg-surface-base/50 border-border-theme-subtle/50',
-  },
-};
-
-/**
- * UI Theme picker
- */
-function UIThemePicker({
-  value,
-  onChange,
-}: {
-  value: UITheme;
-  onChange: (theme: UITheme) => void;
-}) {
-  const themes: UITheme[] = ['default', 'minimal', 'neon', 'tactical'];
-
-  return (
-    <div>
-      <div className="text-text-theme-primary mb-2 text-sm font-medium">
-        UI Theme
-      </div>
-      <div className="text-text-theme-secondary mb-3 text-xs">
-        Choose the overall visual style of the interface
-      </div>
-      <div className="grid grid-cols-2 gap-3">
-        {themes.map((theme) => {
-          const info = UI_THEME_INFO[theme];
-          const isSelected = value === theme;
-
-          return (
-            <button
-              key={theme}
-              onClick={() => onChange(theme)}
-              className={`focus:ring-accent focus:ring-offset-surface-base rounded-md border-2 p-3 text-left transition-all focus:ring-2 focus:ring-offset-2 focus:outline-none ${
-                isSelected
-                  ? 'border-accent bg-accent-muted'
-                  : 'border-border-theme hover:border-border-theme-strong bg-surface-raised/30'
-              } `}
-            >
-              {/* Preview bar */}
-              <div
-                className={`mb-2 h-2 w-full rounded border ${info.preview}`}
-              />
-              <div className="text-text-theme-primary text-sm font-medium">
-                {info.name}
-              </div>
-              <div className="text-text-theme-secondary mt-0.5 text-xs">
-                {info.description}
-              </div>
-            </button>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-// Valid section IDs for type checking
-const VALID_SECTION_IDS: SectionId[] = [
-  'appearance',
-  'customizer',
-  'vault',
-  'p2p-sync',
-  'ui-behavior',
-  'accessibility',
-  'audit',
-  'reset',
-];
-
-function isValidSectionId(hash: string): hash is SectionId {
-  return VALID_SECTION_IDS.includes(hash as SectionId);
-}
-
-/**
- * P2P Sync settings section
- */
-function P2PSyncSection({
-  isExpanded,
-  onToggle,
-  onRef,
-}: {
-  isExpanded: boolean;
-  onToggle: () => void;
-  onRef: (el: HTMLDivElement | null) => void;
-}): React.ReactElement {
-  const connectionState = useConnectionState();
-  const peers = usePeers();
-  const roomCode = useRoomCode();
-  const {
-    localPeerName,
-    localPeerId,
-    setLocalPeerName,
-    createRoom,
-    joinRoom,
-    leaveRoom,
-    error,
-    clearError,
-  } = useSyncRoomStore();
-
-  const [showRoomDialog, setShowRoomDialog] = React.useState(false);
-  const [peerNameInput, setPeerNameInput] = React.useState(localPeerName);
-
-  const isConnected = connectionState === ConnectionState.Connected;
-  const isConnecting = connectionState === ConnectionState.Connecting;
-
-  const handleSavePeerName = () => {
-    if (peerNameInput.trim()) {
-      setLocalPeerName(peerNameInput.trim());
-    }
-  };
-
-  return (
-    <SettingsSection
-      id="p2p-sync"
-      title="P2P Sync"
-      description="Real-time peer-to-peer synchronization for collaborative play"
-      isExpanded={isExpanded}
-      onToggle={onToggle}
-      onRef={onRef}
-    >
-      {/* Connection Status */}
-      <div className="bg-surface-raised/50 border-border-theme-subtle flex items-center justify-between rounded-lg border p-4">
-        <div className="flex items-center gap-4">
-          <SyncStatusIndicator
-            connectionState={connectionState}
-            peerCount={peers.length}
-          />
-          {roomCode && (
-            <div className="text-sm">
-              <span className="text-text-theme-muted">Room: </span>
-              <code className="text-accent font-mono font-bold">
-                {roomCode}
-              </code>
-            </div>
-          )}
-        </div>
-        <div className="flex items-center gap-2">
-          {isConnected ? (
-            <button
-              onClick={leaveRoom}
-              className="rounded-md border border-red-500/30 bg-red-600/20 px-3 py-1.5 text-sm font-medium text-red-400 transition-colors hover:bg-red-600/30"
-            >
-              Disconnect
-            </button>
-          ) : (
-            <button
-              onClick={() => setShowRoomDialog(true)}
-              disabled={isConnecting}
-              className="bg-accent hover:bg-accent-hover rounded-md px-3 py-1.5 text-sm font-medium text-white transition-colors disabled:opacity-50"
-            >
-              {isConnecting ? 'Connecting...' : 'Connect'}
-            </button>
-          )}
-        </div>
-      </div>
-
-      {/* Error Display */}
-      {error && (
-        <div className="flex items-center justify-between rounded-lg border border-red-500/30 bg-red-500/10 p-3">
-          <div className="flex items-center gap-2 text-red-400">
-            <svg
-              className="h-5 w-5"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-              />
-            </svg>
-            <span className="text-sm">{error}</span>
-          </div>
-          <button
-            onClick={clearError}
-            className="text-red-400 hover:text-red-300"
-          >
-            <svg
-              className="h-4 w-4"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M6 18L18 6M6 6l12 12"
-              />
-            </svg>
-          </button>
-        </div>
-      )}
-
-      {/* Peer Name Setting */}
-      <div>
-        <div className="text-text-theme-primary mb-1 text-sm font-medium">
-          Your Display Name
-        </div>
-        <div className="text-text-theme-secondary mb-2 text-xs">
-          How you appear to other peers in sync rooms
-        </div>
-        <div className="flex gap-2">
-          <Input
-            type="text"
-            value={peerNameInput}
-            onChange={(e) => setPeerNameInput(e.target.value)}
-            placeholder="Enter your name"
-            className="flex-1 text-sm"
-          />
-          <button
-            onClick={handleSavePeerName}
-            disabled={!peerNameInput.trim() || peerNameInput === localPeerName}
-            className="bg-accent hover:bg-accent-hover rounded-lg px-4 py-2 text-sm font-medium text-white transition-colors disabled:cursor-not-allowed disabled:opacity-50"
-          >
-            Save
-          </button>
-        </div>
-      </div>
-
-      {/* Connected Peers */}
-      {isConnected && (
-        <div>
-          <div className="text-text-theme-primary mb-2 text-sm font-medium">
-            Connected Peers
-          </div>
-          {peers.length > 0 ? (
-            <PeerList
-              peers={peers}
-              localPeerId={localPeerId}
-              localPeerName={localPeerName}
-            />
-          ) : (
-            <div className="text-text-theme-muted bg-surface-deep/50 border-border-theme-subtle rounded-lg border p-4 text-center text-sm">
-              Waiting for peers to join...
-            </div>
-          )}
-        </div>
-      )}
-
-      {/* How it Works */}
-      <div className="border-border-theme-subtle border-t pt-4">
-        <div className="text-text-theme-primary mb-2 text-sm font-medium">
-          How P2P Sync Works
-        </div>
-        <ul className="text-text-theme-secondary space-y-2 text-xs">
-          <li className="flex items-start gap-2">
-            <span className="text-accent">1.</span>
-            Create a room or join with a room code
-          </li>
-          <li className="flex items-start gap-2">
-            <span className="text-accent">2.</span>
-            Share the room code with your gaming group
-          </li>
-          <li className="flex items-start gap-2">
-            <span className="text-accent">3.</span>
-            Enable sync on vault items you want to share
-          </li>
-          <li className="flex items-start gap-2">
-            <span className="text-accent">4.</span>
-            Changes sync automatically in real-time
-          </li>
-        </ul>
-        <div className="mt-3 rounded-lg border border-amber-500/20 bg-amber-500/10 p-3">
-          <div className="flex items-start gap-2">
-            <svg
-              className="mt-0.5 h-4 w-4 flex-shrink-0 text-amber-400"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-              />
-            </svg>
-            <span className="text-xs text-amber-200">
-              P2P sync uses WebRTC for direct peer connections. All data stays
-              between you and your peers - no server storage.
-            </span>
-          </div>
-        </div>
-      </div>
-
-      {/* Room Dialog */}
-      <RoomCodeDialog
-        isOpen={showRoomDialog}
-        onClose={() => setShowRoomDialog(false)}
-        onCreateRoom={async (password) => {
-          const code = await createRoom(password ? { password } : undefined);
-          return code;
-        }}
-        onJoinRoom={async (code, password) => {
-          await joinRoom(code, password);
-        }}
-        onLeaveRoom={leaveRoom}
-        isConnected={isConnected}
-        isConnecting={isConnecting}
-        currentRoomCode={roomCode}
-        peers={peers}
-        localPeerId={localPeerId}
-        localPeerName={localPeerName}
-        error={error}
-        onClearError={clearError}
-      />
-    </SettingsSection>
-  );
-}
+  AppearanceSettings,
+  CustomizerSettings,
+  VaultSettings,
+  P2PSyncSettings,
+  UIBehaviorSettings,
+  AccessibilitySettings,
+  AuditSettings,
+  ResetSettings,
+  QuickNavigation,
+  isValidSectionId,
+  type SectionId,
+} from '@/components/settings';
+import { PageLayout } from '@/components/ui';
+import { useAppSettingsStore } from '@/stores/useAppSettingsStore';
 
 export default function SettingsPage(): React.ReactElement {
   const router = useRouter();
 
-  // Track hydration to avoid SSR mismatch
   const [_hasMounted, setHasMounted] = useState(false);
-
-  // Active section from URL hash (only one section expanded at a time)
-  // Default to 'appearance' for SSR, will update after mount if hash is present
   const [activeSection, setActiveSection] = useState<SectionId>('appearance');
 
-  // Refs for scrolling to sections
   const sectionRefs = useRef<Record<SectionId, HTMLDivElement | null>>({
     appearance: null,
     customizer: null,
@@ -849,7 +46,6 @@ export default function SettingsPage(): React.ReactElement {
     reset: null,
   });
 
-  // Parse hash from URL on mount and hash changes
   useEffect(() => {
     setHasMounted(true);
 
@@ -857,7 +53,6 @@ export default function SettingsPage(): React.ReactElement {
       const hash = window.location.hash.replace('#', '');
       if (hash && isValidSectionId(hash)) {
         setActiveSection(hash);
-        // Scroll to section after a short delay
         if (shouldScroll) {
           setTimeout(() => {
             const element = sectionRefs.current[hash];
@@ -869,11 +64,9 @@ export default function SettingsPage(): React.ReactElement {
       }
     };
 
-    // Check initial hash (don't scroll immediately on page load to avoid jumpiness)
     const hash = window.location.hash.replace('#', '');
     if (hash && isValidSectionId(hash)) {
       setActiveSection(hash);
-      // Delayed scroll for initial load
       setTimeout(() => {
         const element = sectionRefs.current[hash];
         if (element) {
@@ -882,19 +75,15 @@ export default function SettingsPage(): React.ReactElement {
       }, 150);
     }
 
-    // Listen for hash changes
     const onHashChange = () => handleHashChange(true);
     window.addEventListener('hashchange', onHashChange);
     return () => window.removeEventListener('hashchange', onHashChange);
   }, []);
 
-  // Navigate to section (update URL hash)
   const navigateToSection = useCallback((sectionId: SectionId) => {
-    // Update URL hash (this triggers hashchange event)
     window.history.pushState(null, '', `#${sectionId}`);
     setActiveSection(sectionId);
 
-    // Scroll to section after a short delay
     setTimeout(() => {
       const element = sectionRefs.current[sectionId];
       if (element) {
@@ -903,12 +92,9 @@ export default function SettingsPage(): React.ReactElement {
     }, 50);
   }, []);
 
-  // Toggle section - clicking header expands that section
   const toggleSection = useCallback(
     (sectionId: SectionId) => {
       if (activeSection === sectionId) {
-        // If clicking the active section, keep it open (or optionally close it)
-        // For now, we'll navigate to it anyway to update the URL
         navigateToSection(sectionId);
       } else {
         navigateToSection(sectionId);
@@ -917,55 +103,11 @@ export default function SettingsPage(): React.ReactElement {
     [activeSection, navigateToSection],
   );
 
-  // Get store actions and state
-  const initDraftAppearance = useAppSettingsStore((s) => s.initDraftAppearance);
-  const saveUITheme = useAppSettingsStore((s) => s.saveUITheme);
-  const saveOtherAppearance = useAppSettingsStore((s) => s.saveOtherAppearance);
   const revertAppearance = useAppSettingsStore((s) => s.revertAppearance);
-  const hasUnsavedUITheme = useAppSettingsStore((s) => s.hasUnsavedUITheme);
-  const hasUnsavedOtherAppearance = useAppSettingsStore(
-    (s) => s.hasUnsavedOtherAppearance,
-  );
-
-  // Get draft values for display (or saved values if no draft)
-  const draftAppearance = useAppSettingsStore((s) => s.draftAppearance);
-  const savedAccentColor = useAppSettingsStore((s) => s.accentColor);
-  const savedUITheme = useAppSettingsStore((s) => s.uiTheme);
-  const savedFontSize = useAppSettingsStore((s) => s.fontSize);
-  const savedAnimationLevel = useAppSettingsStore((s) => s.animationLevel);
-  const savedCompactMode = useAppSettingsStore((s) => s.compactMode);
-
-  // Effective values for display
-  const effectiveAccentColor = draftAppearance?.accentColor ?? savedAccentColor;
-  const effectiveUITheme = draftAppearance?.uiTheme ?? savedUITheme;
-  const effectiveFontSize = draftAppearance?.fontSize ?? savedFontSize;
-  const effectiveAnimationLevel =
-    draftAppearance?.animationLevel ?? savedAnimationLevel;
-  const effectiveCompactMode = draftAppearance?.compactMode ?? savedCompactMode;
-
-  // Draft setters for live preview
-  const setDraftAccentColor = useAppSettingsStore((s) => s.setDraftAccentColor);
-  const setDraftUITheme = useAppSettingsStore((s) => s.setDraftUITheme);
-  const setDraftFontSize = useAppSettingsStore((s) => s.setDraftFontSize);
-  const setDraftAnimationLevel = useAppSettingsStore(
-    (s) => s.setDraftAnimationLevel,
-  );
-  const setDraftCompactMode = useAppSettingsStore((s) => s.setDraftCompactMode);
-
-  // Other settings (not draft-based)
-  const settings = useAppSettingsStore();
-
-  // Initialize draft state on mount
-  useEffect(() => {
-    initDraftAppearance();
-  }, [initDraftAppearance]);
-
   const revertCustomizer = useAppSettingsStore((s) => s.revertCustomizer);
 
-  // Revert on unmount if there are unsaved changes
   useEffect(() => {
     const handleRouteChange = () => {
-      // Revert appearance changes when navigating away
       revertAppearance();
       revertCustomizer();
     };
@@ -976,7 +118,6 @@ export default function SettingsPage(): React.ReactElement {
     };
   }, [router.events, revertAppearance, revertCustomizer]);
 
-  // Create ref callback for each section
   const createSectionRef = (id: SectionId) => (el: HTMLDivElement | null) => {
     sectionRefs.current[id] = el;
   };
@@ -992,361 +133,59 @@ export default function SettingsPage(): React.ReactElement {
         subtitle="Customize your MekStation experience"
         maxWidth="narrow"
       >
-        {/* Quick Navigation */}
         <QuickNavigation
           activeSection={activeSection}
           onNavigate={navigateToSection}
         />
 
         <div className="space-y-4 pb-8">
-          {/* Appearance Section */}
-          <SettingsSection
-            id="appearance"
-            title="Appearance"
-            description="Customize colors, fonts, and visual effects. Changes preview instantly but require saving."
+          <AppearanceSettings
             isExpanded={activeSection === 'appearance'}
             onToggle={() => toggleSection('appearance')}
             onRef={createSectionRef('appearance')}
-          >
-            <UIThemePicker
-              value={effectiveUITheme}
-              onChange={setDraftUITheme}
-            />
+          />
 
-            {/* UI Theme Save Notice */}
-            {hasUnsavedUITheme && (
-              <div className="flex items-center justify-between rounded-lg border border-amber-500/30 bg-amber-500/10 p-3">
-                <div className="flex items-center gap-2">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    strokeWidth={1.5}
-                    stroke="currentColor"
-                    className="h-5 w-5 text-amber-400"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      d="M12 9v3.75m9-.75a9 9 0 11-18 0 9 9 0 0118 0zm-9 3.75h.008v.008H12v-.008z"
-                    />
-                  </svg>
-                  <span className="text-sm text-amber-200">
-                    Theme preview active — save to keep changes
-                  </span>
-                </div>
-                <button
-                  onClick={saveUITheme}
-                  className="rounded-md bg-amber-600 px-3 py-1.5 text-sm font-medium text-white transition-colors hover:bg-amber-500"
-                >
-                  Save Theme
-                </button>
-              </div>
-            )}
-
-            <AccentColorPicker
-              value={effectiveAccentColor}
-              onChange={setDraftAccentColor}
-            />
-
-            <Select<FontSize>
-              label="Font Size"
-              description="Base font size for the application"
-              value={effectiveFontSize}
-              onChange={setDraftFontSize}
-              options={[
-                { value: 'small', label: 'Small (14px)' },
-                { value: 'medium', label: 'Medium (16px)' },
-                { value: 'large', label: 'Large (18px)' },
-              ]}
-            />
-
-            <Select<AnimationLevel>
-              label="Animation Level"
-              description="Control the amount of motion and transitions"
-              value={effectiveAnimationLevel}
-              onChange={setDraftAnimationLevel}
-              options={[
-                { value: 'full', label: 'Full - All animations enabled' },
-                {
-                  value: 'reduced',
-                  label: 'Reduced - Essential animations only',
-                },
-                { value: 'none', label: 'None - Disable all animations' },
-              ]}
-            />
-
-            <Toggle
-              label="Compact Mode"
-              description="Reduce spacing and padding for more information density"
-              checked={effectiveCompactMode}
-              onChange={setDraftCompactMode}
-            />
-
-            {/* Save Button - for accent, font, animation, compact (not theme) */}
-            <div className="border-border-theme-subtle border-t pt-4">
-              <div className="flex items-center justify-between">
-                <div className="text-text-theme-secondary text-sm">
-                  {hasUnsavedOtherAppearance
-                    ? 'You have unsaved appearance changes'
-                    : 'Appearance settings saved'}
-                </div>
-                <button
-                  onClick={saveOtherAppearance}
-                  disabled={!hasUnsavedOtherAppearance}
-                  className={`rounded-lg px-4 py-2 text-sm font-medium transition-colors ${
-                    hasUnsavedOtherAppearance
-                      ? 'bg-accent hover:bg-accent-hover text-white'
-                      : 'bg-surface-raised text-text-theme-muted cursor-not-allowed'
-                  }`}
-                >
-                  Save Appearance
-                </button>
-              </div>
-            </div>
-          </SettingsSection>
-
-          {/* Customizer Section */}
-          <SettingsSection
-            id="customizer"
-            title="Customizer"
-            description="Configure the mech customizer interface"
+          <CustomizerSettings
             isExpanded={activeSection === 'customizer'}
             onToggle={() => toggleSection('customizer')}
             onRef={createSectionRef('customizer')}
-          >
-            {/* Armor Diagram Mode */}
-            <div>
-              <div className="text-text-theme-primary mb-2 text-sm font-medium">
-                Armor Diagram Mode
-              </div>
-              <div className="text-text-theme-secondary mb-3 text-xs">
-                Choose between schematic grid or silhouette SVG display
-              </div>
-              <ArmorDiagramModePreview
-                selectedMode={settings.armorDiagramMode}
-                onSelectMode={settings.setArmorDiagramMode}
-              />
-            </div>
+          />
 
-            {/* Armor Diagram Variant (only visible for silhouette mode) */}
-            {settings.armorDiagramMode === 'silhouette' && (
-              <div>
-                <div className="text-text-theme-primary mb-2 text-sm font-medium">
-                  Silhouette Aesthetic
-                </div>
-                <div className="text-text-theme-secondary mb-3 text-xs">
-                  Visual style for the armor diagram
-                </div>
-                <ArmorDiagramSettings />
-              </div>
-            )}
-
-            <Toggle
-              label="Show Design Selector (UAT)"
-              description="Display the design variant dropdown in the armor tab for testing"
-              checked={settings.showArmorDiagramSelector}
-              onChange={settings.setShowArmorDiagramSelector}
-            />
-          </SettingsSection>
-
-          {/* Vault & Sharing Section */}
-          <SettingsSection
-            id="vault"
-            title="Vault & Sharing"
-            description="Manage your vault identity for sharing content"
+          <VaultSettings
             isExpanded={activeSection === 'vault'}
             onToggle={() => toggleSection('vault')}
             onRef={createSectionRef('vault')}
-          >
-            <VaultIdentitySection />
-          </SettingsSection>
+          />
 
-          {/* P2P Sync Section */}
-          <P2PSyncSection
+          <P2PSyncSettings
             isExpanded={activeSection === 'p2p-sync'}
             onToggle={() => toggleSection('p2p-sync')}
             onRef={createSectionRef('p2p-sync')}
           />
 
-          {/* UI Behavior Section */}
-          <SettingsSection
-            id="ui-behavior"
-            title="UI Behavior"
-            description="Control how the interface behaves"
+          <UIBehaviorSettings
             isExpanded={activeSection === 'ui-behavior'}
             onToggle={() => toggleSection('ui-behavior')}
             onRef={createSectionRef('ui-behavior')}
-          >
-            <Toggle
-              label="Collapse Sidebar by Default"
-              description="Start with the sidebar in collapsed state"
-              checked={settings.sidebarDefaultCollapsed}
-              onChange={settings.setSidebarDefaultCollapsed}
-            />
+          />
 
-            <Toggle
-              label="Confirm Before Closing"
-              description="Show a confirmation when closing tabs with unsaved changes"
-              checked={settings.confirmOnClose}
-              onChange={settings.setConfirmOnClose}
-            />
-
-            <Toggle
-              label="Show Tooltips"
-              description="Display helpful tooltips on hover"
-              checked={settings.showTooltips}
-              onChange={settings.setShowTooltips}
-            />
-          </SettingsSection>
-
-          {/* Accessibility Section */}
-          <SettingsSection
-            id="accessibility"
-            title="Accessibility"
-            description="Options for better accessibility"
+          <AccessibilitySettings
             isExpanded={activeSection === 'accessibility'}
             onToggle={() => toggleSection('accessibility')}
             onRef={createSectionRef('accessibility')}
-          >
-            <Toggle
-              label="High Contrast"
-              description="Increase contrast for better visibility"
-              checked={settings.highContrast}
-              onChange={settings.setHighContrast}
-            />
+          />
 
-            <Toggle
-              label="Reduce Motion"
-              description="Minimize animations for motion sensitivity"
-              checked={settings.reduceMotion}
-              onChange={settings.setReduceMotion}
-            />
-          </SettingsSection>
-
-          {/* Audit Log Section */}
-          <SettingsSection
-            id="audit"
-            title="Audit Log"
-            description="View event history and system logs"
+          <AuditSettings
             isExpanded={activeSection === 'audit'}
             onToggle={() => toggleSection('audit')}
             onRef={createSectionRef('audit')}
-          >
-            <div className="space-y-4">
-              <p className="text-text-theme-secondary text-sm">
-                Browse the full event history of your campaigns, pilots, and
-                games. Track changes, review decisions, and replay past
-                sessions.
-              </p>
+          />
 
-              <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-                <Link
-                  href="/audit/timeline"
-                  className="bg-surface-raised/50 border-border-theme-subtle hover:bg-surface-raised hover:border-border-theme group flex items-center gap-3 rounded-lg border p-4 transition-all"
-                >
-                  <div className="bg-accent/20 text-accent group-hover:bg-accent flex h-10 w-10 items-center justify-center rounded-lg transition-colors group-hover:text-white">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      strokeWidth={1.5}
-                      stroke="currentColor"
-                      className="h-5 w-5"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"
-                      />
-                    </svg>
-                  </div>
-                  <div>
-                    <div className="text-text-theme-primary group-hover:text-accent text-sm font-medium transition-colors">
-                      Event Timeline
-                    </div>
-                    <div className="text-text-theme-muted text-xs">
-                      Browse all events
-                    </div>
-                  </div>
-                </Link>
-
-                <div className="bg-surface-raised/30 border-border-theme-subtle/50 flex items-center gap-3 rounded-lg border p-4 opacity-60">
-                  <div className="bg-surface-raised text-text-theme-muted flex h-10 w-10 items-center justify-center rounded-lg">
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      strokeWidth={1.5}
-                      stroke="currentColor"
-                      className="h-5 w-5"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        d="M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-                      />
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        d="M9 9.563C9 9.252 9.252 9 9.563 9h4.874c.311 0 .563.252.563.563v4.874c0 .311-.252.563-.563.563H9.564A.562.562 0 019 14.437V9.564z"
-                      />
-                    </svg>
-                  </div>
-                  <div>
-                    <div className="text-text-theme-secondary text-sm font-medium">
-                      Replay Player
-                    </div>
-                    <div className="text-text-theme-muted text-xs">
-                      Coming soon
-                    </div>
-                  </div>
-                </div>
-              </div>
-
-              <div className="border-border-theme-subtle border-t pt-4">
-                <div className="text-text-theme-muted text-xs">
-                  <strong className="text-text-theme-secondary">Tip:</strong>{' '}
-                  You can also access event history from individual pilot and
-                  force detail pages.
-                </div>
-              </div>
-            </div>
-          </SettingsSection>
-
-          {/* Reset Section */}
-          <SettingsSection
-            id="reset"
-            title="Reset"
+          <ResetSettings
             isExpanded={activeSection === 'reset'}
             onToggle={() => toggleSection('reset')}
             onRef={createSectionRef('reset')}
-          >
-            <div className="flex items-center justify-between">
-              <div>
-                <div className="text-sm font-medium text-white">
-                  Reset All Settings
-                </div>
-                <div className="mt-0.5 text-xs text-slate-400">
-                  Restore all settings to their default values
-                </div>
-              </div>
-              <button
-                onClick={() => {
-                  if (
-                    confirm(
-                      'Are you sure you want to reset all settings to defaults?',
-                    )
-                  ) {
-                    settings.resetToDefaults();
-                  }
-                }}
-                className="rounded-lg bg-red-600 px-4 py-2 text-sm font-medium text-white transition-colors hover:bg-red-500"
-              >
-                Reset
-              </button>
-            </div>
-          </SettingsSection>
+          />
         </div>
       </PageLayout>
     </>
